requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
aiosqlite==0.20.0
alembic==1.13.3
amqp==5.2.0
annotated-types==0.7.0
anyio==4.6.0
argparse==1.4.0
asttokens==2.4.1
asyncpg==0.29.0
attrs==24.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
boto3==1.35.39
botocore==1.35.39
bs4==0.0.2
build==1.2.2.post1
CacheControl==0.14.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cleo==2.1.0
click==8.1.7
configparser==7.1.0
coverage==7.6.2
crashtest==0.4.1
crypto==1.4.1
cryptography==43.0.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
distlib==0.3.9
dnspython==2.7.0
docopt==0.6.2
dulwich==0.21.*
ecdsa==0.19.0
email_validator==2.2.0
executing==2.1.0
fastapi==0.115.0
fastapi-cache2==0.2.2
fastapi-cli==0.0.5
fastjsonschema==2.20.0
filelock==3.12.2
flake8==7.1.1
frozendict==2.4.5
frozenlist==1.4.1
gevent==24.10.2
googlefinance==0.7
greenlet==3.1.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.6
httptools==0.6.1
httpx==0.27.2
idna==3.10
importlib_resources==6.4.5
iniconfig==2.0.0
installer==0.7.0
ipython==8.12.3
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.1
jeepney==0.8.0
Jinja2==3.1.4
jmespath==1.0.1
jose==1.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
keyring==24.*
lazy-object-proxy==1.10.0
limits==3.13.0
lxml==5.3.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==3.0.1
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mistune==3.0.2
more-itertools==10.5.0
msgpack==1.1.0
multidict==6.1.0
multitasking==0.0.11
Naked==0.1.32
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parse==1.20.2
parso==0.8.4
passlib==1.7.4
peewee==3.17.6
pendulum==3.0.0
pexpect==4.9.0
pickleshare==0.7.5
pipreqs==0.5.0
pkginfo==1.11.2
platformdirs==4.3.6
pluggy==1.5.0
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
prompt_toolkit==3.0.48
propcache==0.2.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pycodestyle==2.12.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pyflakes==3.2.0
Pygments==2.18.0
PyMySQL==1.1.1
pyproject_hooks==1.2.0
PySocks==1.7.1
pytest==8.3.3
pytest-cov==5.0.0
pytest-sugar==1.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
RapidFuzz==3.10.0
redis==5.1.1
referencing==0.35.1
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.2
rpds-py==0.20.0
rsa==4.9
safety==3.2.8
s3transfer==0.10.3
SecretStorage==3.3.3
selenium==4.25.0
setuptools==75.1.0
shellescape==3.8.1
shellingham==1.5.4
six==1.16.0
slowapi==0.1.9
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.35
stack-data==0.6.3
starlette==0.*
termcolor==2.5.0
time-machine==2.16.0
tinycss2==1.3.0
tomlkit==0.13.2
tornado==6.4.1
traitlets==5.14.3
trio==0.26.2
trio-websocket==0.11.1
trove-classifiers==2024.10.11
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
undetected-chromedriver==3.5.5
urllib3==2.2.3
uvicorn==0.31.1
uvloop==0.20.0
validator==0.7.1
vine==5.1.0
virtualenv==20.26.6
watchfiles==0.21.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.1
wrapt==1.16.0
wsproto==1.2.0
yarg==0.1.9
yarl==1.14.0
yfinance==0.2.44
zope.event==5.0
zope.interface==7.1.0
asynctest==0.13.0